feat(android): complete Android port — WebView shell + NDK casu_core + widget-ready
- Gradle project (android/): minSdk 24, arm64-v8a/armeabi-v7a/x86_64,
  signed release APK 5.1 MB (keystore excluded from git)
- libcasucore.so via NDK: the byte-parity core sources compile for
  Android; JNI bridge CasuCore.detectKind/verifyCasunat2/extractToCache
- LoopbackServer: same-origin serving of the pure-web touch UI from
  APK assets + /api/version (5.0.0-android)
- MainActivity: WebView with JS/DOM-storage/media-autoplay, loads
  http://127.0.0.1:<port>/web/index.html
- journal.cpp int64_t cast fix for libc++ portability
- Widget + MediaSession: documented as next step in HANDOVER §1.P2

+++ win-release/src/core/casu/journal.cpp
@@ -82,7 +82,7 @@ void write_journal(const std::string& path, const std::string& state,
     JsonObject payload;
     payload.items["version"] = JsonValue(int64_t(1));
     payload.items["state"] = JsonValue(state);
-    payload.items["updated_ns"] = JsonValue(updated_ns);
+    payload.items["updated_ns"] = JsonValue(static_cast<int64_t>(updated_ns));
     auto recorded_jobs = std::make_shared<JsonArray>();
     for (const JournalJob& job : jobs) recorded_jobs->items.push_back(job_record(job));
     payload.items["jobs"] = JsonValue(std::move(recorded_jobs));
